Components on client moved to directories for each type
