+++ client/src/components/model/SingleResident.jsx
