+++ client/src/components/ui/Landing.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
-import ResidentFinder from "./ResidentFinder";
-import AddResident from "./AddResident";
-import { AllResidents, ResidentSearchResult } from "./Residents";
+import ResidentFinder from "../forms/ResidentFinder";
+import AddResident from "../forms/AddResident";
+import { AllResidents, ResidentSearchResult } from "../containers/Residents";
 
 const Landing = () => {
   return (
